Add TCRT documentation and support for it in flashtool
Add documentation for the TCRT file format, which is intended to store
the full memory state of a tapecart in a file, e.g. as serialisation
format in an emulator or as distribution format for programs that run
from a tapecart. Implement writing and dumping of that format in flashtool.